+++ yuyo/modals.py
@@ -58,6 +58,7 @@ import types
 import typing
 
 import alluka as alluka_
+from alluka import local as alluka_local
 import hikari
 import typing_extensions
 
@@ -236,7 +237,7 @@ class ModalClient:
             If `event_managed` is passed as [True][] when `event_manager` is [None][].
         """
         if alluka is None:
-            alluka = alluka_.Client()
+            alluka = alluka_local.get(default=None) or alluka_.Client()
             self._set_standard_deps(alluka)
 
         self._alluka = alluka