+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The component, modal and reaction clients will now  use the "local" Alluka client
+  returned by [alluka.local.get][] if no client is passed and a local Alluka client is set.
+
 ## [1.21.0] - 2024-04-21
 ### Added
 - Component [StreamExecutor][yuyo.components.StreamExecutor] for streaming over received